git-svn-id: https://alma-svn.aoc.nrao.edu/p2/trunk/ACS@135649 523d945c-050c-4681-91ec-863ad3bb968a

+++ LGPL/CommonSoftware/acssampGUI/src/cl/utfsm/samplingSystemUI/SerializableProperty.java
@@ -13,8 +13,8 @@ public class SerializableProperty implements Serializable {
 	private String samplingTime;
 	private String timeWindow;
 	
-	public long getFrequency() {
-		return Long.parseLong(frequency);
+	public double getFrequency() {
+		return Double.parseDouble(frequency);
 	}
 	public void setFrequency(double d) {
 		this.frequency = String.valueOf(d);
@@ -31,7 +31,6 @@ public class SerializableProperty implements Serializable {
 	public void setTimeWindow(int timeWindow) {
 		this.timeWindow = String.valueOf(timeWindow);
 	}
-	
 	public String getComponent() {
 		return component;
 	}
